fix delete e get

+++ src/components/DeleteEvent.jsx
@@ -1,5 +1,5 @@
 // src/components/DeleteEvent.jsx
-import { useState, useEffect } from 'react';
+import { useEffect, useState } from 'react';
 import axios from 'axios';
 
 const DeleteEvent = () => {
@@ -18,34 +18,34 @@ const DeleteEvent = () => {
   }, []);
 
   const handleDelete = async (id) => {
-    try {
-      await axios.delete(`https://xote-api-development.up.railway.app/xote/delete/${id}`);
-      alert('Evento deletado com sucesso!');
-      setEvents(events.filter((event) => event._id !== id)); // Remove o evento da lista localmente
-    } catch (error) {
-      console.error('Erro ao deletar o evento:', error);
+    const confirmDelete = window.confirm("Você realmente deseja deletar este evento?");
+    if (confirmDelete) {
+      try {
+        await axios.delete(`https://xote-api-development.up.railway.app/xote/delete/${id}`);
+        setEvents(events.filter(event => event._id !== id)); // Atualiza a lista removendo o evento deletado
+        alert('Evento excluído com sucesso!');
+      } catch (error) {
+        console.error('Erro ao excluir o evento:', error);
+      }
     }
   };
 
   return (
     <div>
-      <h1>Deletar Eventos</h1>
-      {events.length > 0 ? (
-        <ul>
-          {events.map((event) => (
-            <li key={event._id}>
-              <h3>{event.title}</h3>
-              <img src={event.image_url} alt={event.title} style={{ width: '100px', height: 'auto' }} />
-              <p>{event.description}</p>
-              <p>{`Data: ${event.date} | Hora: ${event.time}`}</p>
-              <a href={event.localgoogleurl} target="_blank" rel="noopener noreferrer">Ver no Google Maps</a>
-              <button onClick={() => handleDelete(event._id)}>Deletar</button>
-            </li>
-          ))}
-        </ul>
-      ) : (
-        <p>Não há eventos cadastrados.</p>
-      )}
+      <h1>Deletar Evento</h1>
+      <ul>
+        {events.map((event) => (
+          <li key={event._id}>
+            <h2>{event.title}</h2>
+            <p>{event.description}</p>
+            <p><strong>Data:</strong> {event.date}</p>
+            <p><strong>Hora:</strong> {event.time}</p>
+            <p><strong>Tipo:</strong> {event.type}</p>
+            <p><strong>Status:</strong> {event.pay ? 'Pago' : 'Não Pago'}</p>
+            <button onClick={() => handleDelete(event._id)}>Excluir Evento</button>
+          </li>
+        ))}
+      </ul>
     </div>
   );
 };
